requirements.txt
torch
transformers
pillow
opencv-python-headless
ffmpeg-python
numpy
yt-dlp
requests
beautifulsoup4
python-dotenv
schedule
python-telegram-bot
tweepy
Mastodon.py
